pyserver/requirements.txt
websockets
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
motor
motor-stubs